+++ Dev/DataModels/Atdi.DataModels.Sdrns.Server.Entities/Entities/ILinkMeasStation.cs
@@ -10,9 +10,9 @@ namespace Atdi.DataModels.Sdrns.Server.Entities
     [Entity]
     public interface ILinkMeasStation
     {
-        int Id { get; set; }
-        int? MeasTaskId { get; set; }
-        int? StationId { get; set; }
+        long Id { get; set; }
+        long? MeasTaskId { get; set; }
+        long? StationId { get; set; }
         IMeasTask MEASTASK { get; set; }
         IStation STATION { get; set; }
     }
