+++ Dev/DataModels/Atdi.DataModels.Sdrns.Server.Entities/Entities/ILinkSectorFreq.cs
@@ -10,9 +10,9 @@ namespace Atdi.DataModels.Sdrns.Server.Entities
     [Entity]
     public interface ILinkSectorFreq
     {
-        int Id { get; set; }
-        int? SectorFreqId { get; set; }
-        int? SectorId { get; set; }
+        long Id { get; set; }
+        long? SectorFreqId { get; set; }
+        long? SectorId { get; set; }
         ISectorFreq SECTORFREQ { get; set; }
         ISector SECTOR { get; set; }
     }
