+++ Dev/DataModels/Atdi.DataModels.Sdrns.Server.Entities/Entities/ILinkSectorMaskElement.cs
@@ -10,9 +10,9 @@ namespace Atdi.DataModels.Sdrns.Server.Entities
     [Entity]
     public interface ILinkSectorMaskElement
     {
-        int Id { get; set; }
-        int? SectorMaskElementId { get; set; }
-        int? SectorId { get; set; }
+        long Id { get; set; }
+        long? SectorMaskElementId { get; set; }
+        long? SectorId { get; set; }
         ISectorMaskElement SECTORMASKELEMENT { get; set; }
         ISector SECTOR { get; set; }
     }
