+++ Dev/DataModels/Atdi.DataModels.Sdrns.Server.Entities/Entities/IMeasDtParam.cs
@@ -10,7 +10,7 @@ namespace Atdi.DataModels.Sdrns.Server.Entities
     [Entity]
     public interface IMeasDtParam
     {
-        int Id { get; set; }
+        long Id { get; set; }
         string TypeMeasurements { get; set; }
         string DetectType { get; set; }
         double? Rfattenuation { get; set; }
@@ -21,7 +21,7 @@ namespace Atdi.DataModels.Sdrns.Server.Entities
         string Mode { get; set; }
         double? Rbw { get; set; }
         double? Vbw { get; set; }
-        int? MeasTaskId { get; set; }
+        long? MeasTaskId { get; set; }
         IMeasTask MEASTASK { get; set; }
     }
 }
