+++ Dev/DataModels/Atdi.DataModels.Sdrns.Server.Entities/Entities/IMeasSubTask.cs
@@ -10,12 +10,12 @@ namespace Atdi.DataModels.Sdrns.Server.Entities
     [Entity]
     public interface IMeasSubTask
     {
-        int Id { get; set; }
+        long Id { get; set; }
         DateTime? TimeStart { get; set; }
         DateTime? TimeStop { get; set; }
         string Status { get; set; }
         int? Interval { get; set; }
-        int? MeasTaskId { get; set; }
+        long? MeasTaskId { get; set; }
         IMeasTask MEASTASK { get; set; }
     }
 }
