+++ Dev/DataModels/Atdi.DataModels.Sdrns.Server.Entities/Entities/IMeasSubTaskSta.cs
@@ -10,12 +10,12 @@ namespace Atdi.DataModels.Sdrns.Server.Entities
     [Entity]
     public interface IMeasSubTaskSta
     {
-        int Id { get; set; }
+        long Id { get; set; }
         string Status { get; set; }
         int? Count { get; set; }
         DateTime? TimeNextTask { get; set; }
-        int? SensorId { get; set; }
-        int? MeasSubTaskId { get; set; }
+        long? SensorId { get; set; }
+        long? MeasSubTaskId { get; set; }
         ISector SENSOR { get; set; }
         IMeasSubTask MEASSUBTASK { get; set; }
     }
