+++ Dev/DataModels/Atdi.DataModels.Sdrns.Server.Entities/Entities/IReferenceLevels.cs
@@ -10,11 +10,11 @@ namespace Atdi.DataModels.Sdrns.Server.Entities
     [Entity]
     public interface IReferenceLevels
     {
-        int Id { get; set; }
+        long Id { get; set; }
         double? StartFrequency_Hz { get; set; }
         double? StepFrequency_Hz { get; set; }
         byte[] ReferenceLevels { get; set; }
-        int? ResMeasId { get; set; }
+        long? ResMeasId { get; set; }
         IResMeas RESMEAS { get; set; }
     }
 }
