+++ Dev/DataModels/Atdi.DataModels.Sdrns.Server.Entities/Entities/IResLocSensorMeas.cs
@@ -10,12 +10,12 @@ namespace Atdi.DataModels.Sdrns.Server.Entities
     [Entity]
     public interface IResLocSensorMeas
     {
-        int Id { get; set; }
+        long Id { get; set; }
         double? Lon { get; set; }
         double? Lat { get; set; }
         double? Asl { get; set; }
         double? Agl { get; set; }
-        int? ResMeasId { get; set; }
+        long? ResMeasId { get; set; }
         IResMeas RESMEAS { get; set; }
     }
 }
