+++ Dev/DataModels/Atdi.DataModels.Sdrns.Server.Entities/Entities/IResStMaskElement.cs
@@ -10,8 +10,8 @@ namespace Atdi.DataModels.Sdrns.Server.Entities
     [Entity]
     public interface IResStMaskElement
     {
-        int Id { get; set; }
-        int? ResStGeneralId { get; set; }
+        long Id { get; set; }
+        long? ResStGeneralId { get; set; }
         double? Bw { get; set; }
         double? Level { get; set; }
         IResStGeneral RESSTGENERAL { get; set; }
