+++ Dev/DataModels/Atdi.DataModels.Sdrns.Server.Entities/Entities/ISector.cs
@@ -10,14 +10,14 @@ namespace Atdi.DataModels.Sdrns.Server.Entities
     [Entity]
     public interface ISector
     {
-        int Id { get; set; }
+        long Id { get; set; }
         double? Agl { get; set; }
         double? Eirp { get; set; }
         double? Azimut { get; set; }
         double? Bw { get; set; }
         string ClassEmission { get; set; }
-        int? StationId { get; set; }
-        int? SectorId { get; set; }
+        long? StationId { get; set; }
+        long? SectorId { get; set; }
         IStation STATION { get; set; }
     }
 }
