+++ Dev/DataModels/Atdi.DataModels.Sdrns.Server.Entities/Entities/IStation.cs
@@ -10,19 +10,19 @@ namespace Atdi.DataModels.Sdrns.Server.Entities
     [Entity]
     public interface IStation
     {
-        int Id { get; set; }
+        long Id { get; set; }
         string GlobalSID { get; set; }
         string Status { get; set; }
         string Standart { get; set; }
-        int? MeasTaskId { get; set; }
-        int? StationId { get; set; }
+        long? MeasTaskId { get; set; }
+        long? StationId { get; set; }
         int? IdPermission { get; set; }
         DateTime StartDate { get; set; }
         DateTime EndDate { get; set; }
         DateTime CloseDate { get; set; }
         string DozvilName { get; set; }
-        int? OwnerDataId { get; set; }
-        int? StationSiteId { get; set; }
+        long? OwnerDataId { get; set; }
+        long? StationSiteId { get; set; }
         IMeasTask MEASTASK { get; set; }
         IStationSite STATIONSITE { get; set; }
         IOwnerData OWNERDATA { get; set; }
