+++ Dev/Modules/Atdi.Modules.Sdrn.Server.Events/OnMeasTaskEvent.cs
@@ -20,8 +20,8 @@ namespace Atdi.Modules.Sdrn.Server.Events
         {
         }
 
-        public int MeasTaskId { get; set; }
-        public int SensorId { get; set; }
+        public long MeasTaskId { get; set; }
+        public long SensorId { get; set; }
         public string MeasTaskIds { get; set; }
         public string SensorName { get; set; }
         public string EquipmentTechId { get; set; }
