+++ Dev/WcfServices/Atdi.WcfServices.Sdrn.Server/Handlers/Analitics/LevelMeasurementsCarForSO.cs
@@ -9,8 +9,8 @@ namespace Atdi.WcfServices.Sdrn.Server
         public DateTime? TimeOfMeasurements; // из class LevelMeasurementsCar
         public decimal? CentralFrequency; // из class LevelMeasurementsCar
         public double? BW; // из class LevelMeasurementsCar если значение <= 0, то вытаскивать из MeasurementsParameterGeneral причем = SpecrumSteps*(T2-T1);
-        public int? Idstation; // из class ResultsMeasurementsStation
-        public int? Id; // из class ResultsMeasurementsStation
+        public long? Idstation; // из class ResultsMeasurementsStation
+        public long? Id; // из class ResultsMeasurementsStation
         public string globalSid;
     }
